+++ Internal/Scripts/Core/Command.cs
@@ -1,8 +1,11 @@
-﻿namespace MobileConsole
+﻿using System;
+
+namespace MobileConsole
 {
 	public class Command
 	{
 		public CommandInfo info = new CommandInfo();
+		public Action refreshUI;
 
 		public virtual void InitDefaultVariableValue() {}