+++ Internal/Scripts/Core/CommandInfo.cs
@@ -11,7 +11,7 @@ namespace MobileConsole
 		public string name;
 		public string description;
 		public int order;
-		public bool shouldCloseAfterExecuted = true;
+		public ActionAfterExecuted actionAfterExecuted = ActionAfterExecuted.CloseAllSubView;
 
 		public bool IsComplex()
 		{