+++ Scripts/DefaultCommands/PersistentData/PersistentDataViewBuilder.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using MobileConsole.UI;
 using UnityEngine;
 
@@ -12,13 +14,23 @@ namespace MobileConsole
 		ImageInspectorViewBuilder _imageInspectorViewBuilder = new ImageInspectorViewBuilder();
 		NodeView _resultNode;
 		string[] _filePaths;
+		bool _isSelectableMode;
+		readonly HashSet<FileInfo> _selectedFiles = new HashSet<FileInfo>(new FileInfoComparer());
+
 
 		public PersistentDataViewBuilder()
 		{
 			Node actionGroup = CreateCategories("Actions");
 			AddButton("Refresh", "action", RefreshLocalFiles, actionGroup);
-			AddButton("Delete All Files", "delete", DeleteAllFiles, actionGroup);
-			AddButton("Share All Files", "share", ShareAllFiles, actionGroup);
+
+			Node allFilesGroup = CreateCategories("All Files", actionGroup);
+			AddButton("Delete All Files", "delete", DeleteAllFiles, allFilesGroup);
+			AddButton("Share All Files", "share", ShareAllFiles, allFilesGroup);
+
+			var selectedGroup = CreateCategories("Selected", actionGroup);
+			AddCheckbox("Selectable Mode", false, OnSelectableModeChanged, selectedGroup);
+			AddButton("Delete Selected Files", "delete", DeleteSelectedFiles, selectedGroup);
+			AddButton("Share Selected Files", "share", ShareSelectedFiles, selectedGroup);
 		}
 
 		public override void OnPrepareToShow()
@@ -58,8 +70,41 @@ namespace MobileConsole
 
 				FileInfo fileInfo = new FileInfo(path);
 				string nodeName = string.Format("{0} <alpha=#66>({1:0.#} KB)", Path.GetFileName(path), fileInfo.Length / 1024.0f);
-				Node fileNode = AddButton(nodeName, InspectFile, categoryNode);
-				fileNode.data = fileInfo;
+				if (_isSelectableMode)
+				{
+					bool isOn = _selectedFiles.Contains(fileInfo);
+					// if (_selectedFilePaths.TryGetValue(fileInfo, out var existedValue))
+					// {
+					// 	isOn = true;
+					// 	fileInfo = existedValue;
+					// }
+					Node fileNode = AddCheckbox(nodeName, isOn, OnFileSelectChanged, categoryNode);
+					fileNode.data = fileInfo;
+				}
+				else
+				{
+					Node fileNode = AddButton(nodeName, InspectFile, categoryNode);
+					fileNode.data = fileInfo;
+				}
+			}
+		}
+
+		private void OnSelectableModeChanged(CheckboxNodeView node)
+		{
+			_isSelectableMode = node.isOn;
+			RefreshLocalFiles(null);
+		}
+
+		private void OnFileSelectChanged(CheckboxNodeView node)
+		{
+			FileInfo fileInfo = (FileInfo)node.data;
+			if (node.isOn)
+			{
+				_selectedFiles.Add(fileInfo);
+			}
+			else
+			{
+				_selectedFiles.Remove(fileInfo);
 			}
 		}
 
@@ -105,11 +150,21 @@ namespace MobileConsole
 			RefreshLocalFiles(null);
 		}
 
+		void DeleteSelectedFiles(GenericNodeView nodeView)
+		{
+			foreach (var fileInfo in _selectedFiles)
+			{
+				fileInfo.Delete();
+			}
+			_selectedFiles.Clear();
+			RefreshLocalFiles(null);
+		}
+
 		void ShareAllFiles(GenericNodeView nodeView)
 		{
 			if (_filePaths == null)
 			{
-				_filePaths = Directory.GetFiles(Application.persistentDataPath, "*", SearchOption.AllDirectories);	
+				_filePaths = Directory.GetFiles(Application.persistentDataPath, "*", SearchOption.AllDirectories);
 			}
 
 			if (_filePaths.Length > 0)
@@ -118,10 +173,35 @@ namespace MobileConsole
 			}
 		}
 
+		void ShareSelectedFiles(GenericNodeView nodeView)
+		{
+			if (_selectedFiles.Count > 0)
+			{
+				string[] filePaths = _selectedFiles.Select(f => f.FullName).ToArray();
+				NativeShare.ShareMultiple("", filePaths);
+			}
+		}
+
 		protected override void OnCategoryToggled(GenericNodeView nodeView)
 		{
 			base.OnCategoryToggled(nodeView);
 			CategoryPlayerPrefs.SaveCategoryState(nodeView, ClassName);
+		}
+
+		private class FileInfoComparer : IEqualityComparer<FileInfo>
+		{
+			public bool Equals(FileInfo x, FileInfo y)
+			{
+				if (x == null || y == null)
+					return false;
+
+				return StringComparer.Ordinal.Equals(x.FullName, y.FullName);
+			}
+
+			public int GetHashCode(FileInfo obj)
+			{
+				return StringComparer.Ordinal.GetHashCode(obj.FullName);
+			}
 		}
 	}
 }